+++ src/migrations/Farm.Migrations.PostgreSQL/Migrations/20260317045807_AddBedPreConfirmed.cs
@@ -1,0 +1,29 @@
+﻿using Microsoft.EntityFrameworkCore.Migrations;
+
+#nullable disable
+
+namespace Farm.Migrations.PostgreSQL.Migrations
+{
+    /// <inheritdoc />
+    public partial class AddBedPreConfirmed : Migration
+    {
+        /// <inheritdoc />
+        protected override void Up(MigrationBuilder migrationBuilder)
+        {
+            migrationBuilder.AddColumn<bool>(
+                name: "BedPreConfirmed",
+                table: "Printers",
+                type: "boolean",
+                nullable: false,
+                defaultValue: false);
+        }
+
+        /// <inheritdoc />
+        protected override void Down(MigrationBuilder migrationBuilder)
+        {
+            migrationBuilder.DropColumn(
+                name: "BedPreConfirmed",
+                table: "Printers");
+        }
+    }
+}